+++ examples/be-example-express-sqlite/CHANGELOG.md
@@ -1,5 +1,13 @@
 # be-example-express-sqlite
 
+## 0.0.11
+
+### Patch Changes
+
+- Update prisma clis
+- Updated dependencies
+  - convostack@0.0.13
+
 ## 0.0.10
 
 ### Patch Changes